fix: presentational에 상태관리 로직을 두었던 부분 수정
