+++ src/presentationals/FormulaRepresentation.jsx
@@ -1,10 +1,6 @@
 import React from "react";
-import { useSelector, useDispatch } from "react-redux";
 import styled from "styled-components";
 import { addStyles, EditableMathField } from "react-mathquill";
-
-import { setLatexInput } from "../slice";
-import { latexFunction } from "../util";
 
 addStyles();
 
@@ -15,26 +11,13 @@ const FormulaRepresentationStyle = styled.div`
 	flex-direction: row;
 `;
 
-export default function FormulaRepresentation() {
-	const dispatch = useDispatch();
-	const latexInput = useSelector(state => state.latexInput);
-
-	const handleLatexInput = mathField => {
-		dispatch(setLatexInput(mathField.latex()));
-	};
-
-	const setUpLatexInsertFunction = mathField => {
-		latexFunction.insertLatex = latex => {
-			mathField.write(latex);
-		};
-	};
-
+export default function FormulaRepresentation({ latexInput, handleLatexInput, mathquillDidMount }) {
 	return (
 		<FormulaRepresentationStyle>
 			<EditableMathField
 				latex={latexInput}
 				onChange={handleLatexInput}
-				mathquillDidMount={setUpLatexInsertFunction}
+				mathquillDidMount={mathquillDidMount}
 			/>
 		</FormulaRepresentationStyle>
 	);
